+++ packages/orq-rc/RELEASES.md
@@ -10679,3 +10679,13 @@ Based on:
 - [python v4.15.0-rc.20] packages/orq-rc
 ### Releases
 - [PyPI v4.15.0-rc.20] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.20 - packages/orq-rc
+
+## 2026-09-04 08:06:37
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.796.4 (2.935.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.15.0-rc.21] packages/orq-rc
+### Releases
+- [PyPI v4.15.0-rc.21] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.21 - packages/orq-rc